#479 Added xUnit as a unit testing framework and restructured existing tests. On top of that, added Effort (and EF6) packages to query tests project.
