+++ Whoaverse/Tests/Query.Tests/packages.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Effort.EF6" version="1.1.4" targetFramework="net45" />
+  <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
+  <package id="NMemory" version="1.1.0" targetFramework="net45" />
+  <package id="xunit" version="2.0.0" targetFramework="net45" />
+  <package id="xunit.abstractions" version="2.0.0" targetFramework="net45" />
+  <package id="xunit.assert" version="2.0.0" targetFramework="net45" />
+  <package id="xunit.core" version="2.0.0" targetFramework="net45" />
+  <package id="xunit.extensibility.core" version="2.0.0" targetFramework="net45" />
+</packages>
