--- Whoaverse/UnitTests/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="HtmlAgilityPack" version="1.4.9" targetFramework="net45" />
-</packages>
